pyproject.toml
[tool.poetry]
name = "sentinel"
version = "0.1.0"
description = ""
authors = ["Paul Gessinger <hello@paulgessinger.com>"]
packages = [
    { include = "sentinel", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
gidgethub = "^5.1.0"
aiohttp = "^3.8.1"
python-dotenv = "^0.19.2"
pysmee = "^0.1.1"
sanic = "^21.12.1"
cachetools = "^5.0.0"
python-dateutil = "^2.8.2"
pydantic = "^1.9.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
